pyproject.toml
[project]
name = "ble-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bleak>=3.0.1",
    "paho-mqtt>=2.1.0",
    "pydantic>=2.13.3",
]
